make the value, and not the character, be logged on unusual unicode found in XmlWriter

+++ src/Xml/XmlWriter.cxx
@@ -425,7 +425,7 @@ XmlWriter::writeCharacters(const AString& text) throw(XmlException) {
          writer << c;
       }
       else {
-          CaretLogWarning("Unicode value of character not written: " + c);
+          CaretLogWarning("Unicode value of character not written: " + (int)c);
       }
    }
    //writer << (text);
